Issue #73: explicitly ignore 3xx HTTP errors (#74)
* Add test assertion that ping seq# increments correctly
* Issue #73: explicitly ignore 3xx HTTP errors
Issue #73 is not a bug. HTTP code 3xx errors never bubble up to the TelemetryScheduler
as they are internally handled by URLSession (and may get reported as other
forms of URL connection errors). The existing code is correct.
However, it doesn't hurt to be explicit here that we would not delete the ping
if 3xx error happens.